Refactor labeling application structure and enhance configuration
- Updated the configuration file to include new color settings and improved organization of constants.
- Introduced a new MessageHandler class to manage communication with the VS Code extension, enhancing modularity.
- Created CanvasManager and UIManager classes to encapsulate canvas and UI-related functionalities, improving code maintainability.
- Refactored LabelingState to streamline state management and improve initialization logic.
- Added event handling for keyboard shortcuts and improved error handling in the application.
- Enhanced the labeling panel's HTML structure for better accessibility and user interaction.
