+++ src/templates/config.js
@@ -1,25 +1,48 @@
-// Constants and Configuration
+/**
+ * 应用程序配置常量
+ */
 export const CONFIG = {
-    CLOSE_POINT_THRESHOLD: 0.01,
+    // 颜色设置
     COLORS: [
-        '#2196F3', '#4CAF50', '#F44336', '#FFC107', '#9C27B0',
-        '#00BCD4', '#FF9800', '#795548', '#607D8B', '#E91E63'
+        '#ff3b30', // 红
+        '#4cd964', // 绿
+        '#007aff', // 蓝
+        '#ff9500', // 橙
+        '#5856d6', // 紫
+        '#ffcc00', // 黄
+        '#34c759', // 草绿
+        '#ff2d55', // 粉红
+        '#5ac8fa', // 浅蓝
+        '#af52de'  // 紫红
     ],
-    DEBOUNCE_DELAY: 200,
-    MAX_SEARCH_RESULTS: 10,
+    CROSSHAIR_COLOR: 'rgba(255, 255, 255, 0.5)',
+    CROSSHAIR_CENTER_COLOR: '#ff3b30',
+    BACKGROUND_COLOR: '#1e1e1e', // 默认背景色，会被VS Code主题覆盖
+
+    // 绘图设置
+    LINE_WIDTH: 2,
+    POINT_RADIUS: 5,
+    HIGHLIGHT_RADIUS: 8,
+    CLOSE_HIGHLIGHT_RADIUS: 12,
+    LABEL_HEIGHT: 20,
+    LABEL_FONT_SIZE: 14,
+    LABEL_PADDING: 5,
+    
+    // 缩放设置
     MIN_ZOOM: 0.1,
     MAX_ZOOM: 10,
     ZOOM_SPEED: 0.1,
-    SCROLL_SPEED: 30,
-    POINT_RADIUS: 3,
-    HIGHLIGHT_RADIUS: 5,
-    CLOSE_HIGHLIGHT_RADIUS: 8,
-    MIN_BOX_SIZE: 0.01, // 1% of image size
-    LINE_WIDTH: 2,
-    CROSSHAIR_COLOR: 'rgba(0, 255, 0, 0.94)',
-    CROSSHAIR_CENTER_COLOR: 'rgba(255, 255, 255, 0.9)',
-    BACKGROUND_COLOR: getComputedStyle(document.documentElement).getPropertyValue('--vscode-editor-background') || '#1e1e1e',
-    LABEL_FONT_SIZE: 14,
-    LABEL_PADDING: 5,
-    LABEL_HEIGHT: 20
-}; 
+    SCROLL_SPEED: 20,
+    
+    // 多边形设置
+    CLOSE_POINT_THRESHOLD: 0.02, // 关闭多边形的距离阈值（归一化值）
+    
+    // 边界框设置
+    MIN_BOX_SIZE: 0.01, // 最小边界框大小（归一化值）
+    
+    // 防抖和节流
+    DEBOUNCE_DELAY: 300,
+    
+    // 搜索设置
+    MAX_SEARCH_RESULTS: 10,
+}
