+++ src/templates/labeling-panel.html
@@ -6,7 +6,7 @@
     <title>YOLO Labeling</title>
     <link rel="stylesheet" href="labeling-panel.css">
 </head>
-<body>
+<body tabindex="0">
     <div class="container">
         <!-- 顶部工具栏 -->
         <header class="toolbar">
